Create README.md
An academic discipline or field of study is a branch of knowledge, taught and researched as part of higher education. A scholar’s discipline is commonly defined by the university faculties and learned societies to which they belong and the academic journals in which they publish research. Thus, if referring to an expert in the discipline of ‘disciplinics‘, the correct professional title for that role would be a ‘disciplinist‘.

Disciplines vary between well-established ones that exist in almost all universities and have well-defined rosters of journals and conferences, and nascent ones supported by only a few universities and publications. A discipline may have branches, and these are often called sub-disciplines. The following outline is provided as an overview of and topical guide to academic disciplines. In each case an entry at the highest level of the hierarchy (e.g., Social sciences, Natural sciences, Engineering, Humanities) is a group of broadly similar disciplines; an entry at the next highest level (e.g., Biology) is a discipline having some degree of autonomy and being the basic identity felt by its scholars; and lower levels of the hierarchy are sub-disciplines not normally having any role in the structure of the university’s governance.